Land Enemy is now a separate script from EnemyBase
PENDING: Turn Water Enemy into a separate script as well.
