+++ Assets/Scenes/SampleScene.unity
@@ -292,10 +292,10 @@ PolygonCollider2D:
   m_AutoTiling: 0
   m_Points:
     m_Paths:
-    - - {x: -18.833015, y: 12.964346}
+    - - {x: -18.817648, y: 14.024659}
       - {x: -18.825777, y: -12.9407215}
       - {x: 18.814602, y: -12.9272995}
-      - {x: 18.792131, y: 12.960446}
+      - {x: 18.772982, y: 14.032841}
   m_UseDelaunayMesh: 0
 --- !u!1 &256815822
 GameObject:
