+++ Assets/Scripts/Enemy/Fisherman.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72c23e35090119f48b9749ec0a4496db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
