+++ Assets/Scripts/Enemy/Fisherman/FishermanConfig.cs
