+++ Assets/Scripts/Enemy/LandEnemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d201e1588ec4ca4382866d4bc8d3d39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
