+++ Assets/Scripts/Enemy/LandEnemy/LandEnemyScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9582d563384e26741b9d058e52450333
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
